move the node to old category
