+++ old_nodes/node_remote.py
@@ -29,7 +29,7 @@ from bpy.props import (
 
 from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.data_structure import dataCorrect, updateNode
-from sverchok.nodes.object_nodes.getsetprop import assign_data, types
+from sverchok.old_nodes.getsetprop import assign_data
 
 
 class SvNodePickup(bpy.types.Operator):
